requirements.txt
# Data Collection
polygon-api-client==1.13.4
pandas==2.2.0
numpy==1.26.3
ta-lib==0.4.28

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
torch==2.2.0
torchvision==0.17.0

# API & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Utils
apscheduler==3.10.4
pydantic-settings==2.1.0
loguru==0.7.2
pytz==2024.1
python-dotenv==1.0.0

# Development
pytest==8.0.0
black==24.1.1
flake8==7.0.0
